hybernate renew

+++ YouRecipe/src/main/java/by/bsu/ilyin/service/RecipeService.java
@@ -1,11 +1,11 @@
 package by.bsu.ilyin.service;
 
 import by.bsu.ilyin.dao.RecipeDAO;
-import by.bsu.ilyin.entities.Recipe;
+import by.bsu.ilyin.hibernate.Recipe;
 
 import java.sql.SQLException;
 
-public class RecipeService extends Service<Recipe,Integer> {
+public class RecipeService extends Service<Recipe,Long> {
 
     public RecipeService() throws SQLException, ClassNotFoundException {
         DAO =new RecipeDAO();
